sentiment_analysis_kafka/requirements.txt
openai==0.27.0
kafka-python==2.0.2
python-dotenv==1.0.0
certifi>=2024.8.30
